Less expensive jsonrpc logging (bug#79361)
Remove the oldest 1/4 of the jsonrpc events buffer when reaching the
size limit instead of just a few lines.  This reduces the cost of adding
a log entry from O(buffer-size) to O(1).

Also make messages forwarded to the events buffer, such as ones sent to
stderr from the server process, obey the same limit.

* lisp/jsonrpc.el (jsonrpc--limit-buffer-size): New.
(jsonrpc--log-event, jsonrpc--forwarding-buffer): Use it.